// blockchain/typechain-types/EvidenceRegistry.ts
export declare namespace EvidenceRegistry {
  export type AuditEntryStruct = {
    category: BigNumberish;
    severity: BigNumberish;
    description: string;
    investigator: AddressLike;
    investigationId: string;
    evidenceId: string;
    timestamp: BigNumberish;
    metadata: BytesLike;
    eventHash: BytesLike;
  };

  export type AuditEntryStructOutput = [
    category: bigint,
    severity: bigint,
    description: string,
    investigator: string,
    investigationId: string,
    evidenceId: string,
    timestamp: bigint,
    metadata: string,
    eventHash: string
  ] & {
    category: bigint;
    severity: bigint;
    description: string;
    investigator: string;
    investigationId: string;
    evidenceId: string;
    timestamp: bigint;
    metadata: string;
    eventHash: string;
  };

  export type EvidenceDataStruct = {
    evidenceId: string;
    evidenceHash: BytesLike;
    timestamp: BigNumberish;
    investigator: AddressLike;
    investigationId: string;
    verificationStatus: BigNumberish;
    metadata: BytesLike;
  };

  export type EvidenceDataStructOutput = [
    evidenceId: string,
    evidenceHash: string,
    timestamp: bigint,
    investigator: string,
    investigationId: string,
    verificationStatus: bigint,
    metadata: string
  ] & {
    evidenceId: string;
    evidenceHash: string;
    timestamp: bigint;
    investigator: string;
    investigationId: string;
    verificationStatus: bigint;
    metadata: string;
  };

  export type VerificationRecordStruct = {
    evidenceId: string;
    verifier: AddressLike;
    timestamp: BigNumberish;
    result: boolean;
    expectedHash: BytesLike;
    actualHash: BytesLike;
    status: BigNumberish;
  };

  export type VerificationRecordStructOutput = [
    evidenceId: string,
    verifier: string,
    timestamp: bigint,
    result: boolean,
    expectedHash: string,
    actualHash: string,
    status: bigint
  ] & {
    evidenceId: string;
    verifier: string;
    timestamp: bigint;
    result: boolean;
    expectedHash: string;
    actualHash: string;
    status: bigint;
  };
}